Revise the wrong datatype for pixel coordinate back to INT; Implement readCSV for initialize the list of gameUnits Not total Successfully to display player:directions are ok BUT can not yet get continuous keyboard input

+++ src/Dot.java
@@ -5,7 +5,7 @@ public class Dot extends GameUnit{
     boolean isExist;
     private final static Image dot = new Image("res/dot.png");
 
-    public Dot(double coordinateX, double coordinateY) {
+    public Dot(int coordinateX, int coordinateY) {
         super(coordinateX, coordinateY);
         this.isExist = true;
     }
